+++ include/network_functions/PsiVector.hpp
@@ -9,6 +9,6 @@ using namespace std;
 
 
 template<typename Psi_t>
-void psi_vector(complex<double>* result, const Psi_t& psi);
+void psi_vector(complex<float>* result, const Psi_t& psi);
 
 } // namespace rbm_on_gpu
